+++ qml/CopyableLabel.qml
@@ -3,9 +3,11 @@ import QtQuick 2.12
 import QtQuick.Controls 2.13
 
 Label {
+    property real delay: 500
+    id: self
     background: MouseArea {
         ToolTip.text: qsTrId('id_click_to_copy')
-        ToolTip.delay: 500
+        ToolTip.delay: self.delay
         ToolTip.visible: containsMouse
         hoverEnabled: true
         onClicked: {