+++ qml/SendDialog.qml
@@ -12,6 +12,7 @@ ControllerDialog {
     wallet: send_dialog.account.wallet
 
     controller: SendController {
+        id: send_controller
         account: send_dialog.account
         balance: send_view.balance
         address: send_view.address
@@ -24,5 +25,36 @@ ControllerDialog {
 
     initialItem: SendView {
         id: send_view
+    }
+
+    doneComponent: WizardPage {
+        actions: Action {
+            text: qsTrId('id_ok')
+            onTriggered: send_dialog.accept()
+        }
+        contentItem: ColumnLayout {
+            spacing: constants.p1
+            VSpacer {}
+            Image {
+                Layout.alignment: Qt.AlignHCenter
+                source: 'qrc:/svg/check.svg'
+                sourceSize.width: 64
+                sourceSize.height: 64
+            }
+            Label {
+                Layout.alignment: Qt.AlignHCenter
+                id: doneLabel
+                text: doneText
+                font.pixelSize: 20
+            }
+            CopyableLabel {
+                Layout.alignment: Qt.AlignHCenter
+                Layout.topMargin: constants.p1
+                font.pixelSize: 12
+                delay: 50
+                text: send_controller.signedTransaction.data.txhash
+            }
+            VSpacer {}
+        }
     }
 }