+++ src/controllers/sendcontroller.cpp
@@ -242,11 +242,26 @@ void SendController::signAndSend()
         details["memo"] = m_memo;
         auto send = new SendTransactionHandler(wallet(), details);
         connect(send, &Handler::done, this, [this, send] {
-           send->deleteLater();
-           wallet()->updateConfig();
-           emit finished();
+            auto tx = send->result().value("result").toObject();
+            setSignedTransaction(m_account->getOrCreateTransaction(tx));
+            send->deleteLater();
+            wallet()->updateConfig();
+            emit finished();
         });
         exec(send);
     });
     exec(sign);
+}
+
+Transaction* SendController::signedTransaction() const
+{
+    return m_signed_transaction;
+}
+
+void SendController::setSignedTransaction(Transaction* signed_transaction)
+{
+    if (m_signed_transaction!=signed_transaction) {
+        m_signed_transaction = signed_transaction;
+        emit signedTransactionChanged(m_signed_transaction);
+    }
 }